Handle unknown symbol warning message
